[common.py] keep user-specified headers when faker is True
Previously, for multiple functions, when the faker argument is True and
the headers argument is non-empty, the headers argument is silently
ignored. This commit fixes that behavior by merging fake headers with
user-specified headers (but not overriding user-specified keys).

Also, get rid of mutable values as default arguments, which is an
anti-pattern and not safe for changes like the one above. As part of
this effort, when a default value {} is replaced with None, the
corresponding argument is also copied on use, which may be a bit
paranoid in some cases.