[bilibili] use faker when downloading videos
Bilibili's CDN now may reject requests with 403 when urllib's default
user agent is used. This appears to be location-based (and started today
for me; others might have experienced this for a while): I tested on a
bunch of U.S.-based nodes and all requests were rejected without a
spoofed user agent, but when I tested on a Mainland China-based node,
the requests went through.