Revert argparse, back to getopt
Because "moving from gnu_getopt to argparse brings more problems than it
solves, if it really solves any". It contains "grammar errors brought with
by a new patch", doesn't fix "something defected exsited in the old
code". "there is no reason to blame the old code."

What's more, "argparse is much more complex with a lengthy document", it
makes you-get pythonic, but we need to make "Those who only know js or
as3 can contribute because you-get is not pythonic".

"This patch is premature and personal taste."

If you read the help again and again to find a flag that you know
you-get has but failed, you can just read the source code. And if you're
familiar with "js or as3", probably you can contribute to fix it.
"keeping description up-to-date" isn't a tough task after all.

This reverts commit bcd8d74d73f13f310a92929c5654e208ef5140eb.

This commit closes #2351, #2355 and probably #2145 too.